Move toBeEmptyAssertion to the element assertions
